refactor(liblili2json): clean up __init__.py
