+++ workspaces/liblili2json/liblili2json/__init__.py
@@ -2,7 +2,6 @@
 Export LibreLingo courses in the JSON format expected by the web app
 """
 
-import itertools
 from .types import *
 from .utils import *
 from .dictionary import *
@@ -11,15 +10,3 @@ from .module import *
 from .course import *
 
 __version__ = '0.1.0'
-
-
-def make_challenges_using(callback, data_source, course):
-    return list(itertools.chain(
-        *map(lambda item: callback(item, course), data_source)))
-
-
-def get_challenges_data(skill, course):
-    return sum([
-        make_challenges_using(get_phrase_challenges, skill.phrases, course),
-        make_challenges_using(get_word_challenges, skill.words, course),
-    ], start=[])
